feat: "pleb upgrade" special-cases @types/node (#91)
if @types/node has a numeric-only request (e.g. "12"), it will ignore it and not upgrade. this allows `"@types/node": "10"` to be kept as-is.
while a configuration file could (or should) be introduced, the aim is to avoid repetitive project-specific config and have better default behavior for the existing ecosystem.